fix: adjust Operate after 8.5.0 release
closes #17371
