+++ operate/importer-8_6/pom.xml
@@ -7,9 +7,9 @@
     <version>8.6.0-SNAPSHOT</version>
     <relativePath>../pom.xml</relativePath>
   </parent>
-  <artifactId>operate-importer-8_5</artifactId>
+  <artifactId>operate-importer-8_6</artifactId>
 
-  <name>Operate Importer 8.5</name>
+  <name>Operate Importer 8.6</name>
 
   <dependencies>
 
@@ -40,7 +40,6 @@
     </dependency>
 
     <!-- SPRING -->
-
     <dependency>
       <groupId>org.springframework</groupId>
       <artifactId>spring-context</artifactId>
@@ -118,7 +117,7 @@
           <relocations>
             <relocation>
               <pattern>io.camunda.zeebe.protocol</pattern>
-              <shadedPattern>io.camunda.zeebe.protocol.v850</shadedPattern>
+              <shadedPattern>io.camunda.zeebe.protocol.v860</shadedPattern>
             </relocation>
           </relocations>
         </configuration>
