+++ operate/importer/pom.xml
@@ -29,13 +29,14 @@
 
     <dependency>
       <groupId>io.camunda</groupId>
-      <artifactId>operate-importer-8_4</artifactId>
+      <artifactId>operate-importer-8_5</artifactId>
       <scope>runtime</scope>
     </dependency>
 
     <dependency>
       <groupId>io.camunda</groupId>
-      <artifactId>operate-importer-8_5</artifactId>
+      <artifactId>operate-importer-8_6</artifactId>
+      <version>8.6.0-SNAPSHOT</version>
       <scope>runtime</scope>
     </dependency>
 
@@ -149,7 +150,7 @@
         <artifactId>maven-dependency-plugin</artifactId>
         <configuration>
           <usedDependencies>
-            <dependency>io.camunda:operate-importer-8_4</dependency>
+            <dependency>io.camunda:operate-importer-8_6</dependency>
           </usedDependencies>
         </configuration>
       </plugin>
