+++ operate/qa/migration-tests/migration-test/src/test/resources/test.properties
@@ -1,1 +1,1 @@
-upgrade.path=1.1.0,1.2.0,1.3.0,8.0.18,8.1.14,8.2.8,8.3.1,8.4.0
+upgrade.path=1.1.0,1.2.0,1.3.0,8.0.18,8.1.14,8.2.8,8.3.1,8.4.0,8.5.0
