+++ operate/qa/migration-tests/pom.xml
@@ -21,6 +21,7 @@
     <module>test-fixture-820</module>
     <module>test-fixture-830</module>
     <module>test-fixture-840</module>
+    <module>test-fixture-850</module>
     <module>migration-test</module>
   </modules>
